Move page bodies into features/ and colocate single-route screens.
Learning, exercises, and account are shared domains; recovery stays in account rather than splitting auth from profile. Screens used by one route sit next to their page.tsx. pages-client is gone.
