+++ src/app/(app)/(learn)/talk/page.tsx
@@ -1,4 +1,4 @@
-import Talk from "@/pages-client/Talk";
+import Talk from "./Talk";
 
 export default function Page() {
   return <Talk />;
