+++ src/app/(app)/(public)/page.tsx
@@ -1,4 +1,4 @@
-import Home from "@/pages-client/Home";
+import Home from "./Home";
 
 export default function Page() {
   return <Home />;
