+++ src/app/(app)/(public)/stories/page.tsx
@@ -1,4 +1,4 @@
-import Stories from "@/pages-client/Stories";
+import Stories from "./Stories";
 
 export default function Page() {
   return <Stories />;
