+++ src/features/exercises/components/MatchAudioExercisePlaceholder.tsx
@@ -7,8 +7,8 @@ import GraphicEqRoundedIcon from "@mui/icons-material/GraphicEqRounded";
 import ImageRoundedIcon from "@mui/icons-material/ImageRounded";
 import CheckRoundedIcon from "@mui/icons-material/CheckRounded";
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
-import type { ChoiceCandidate } from "../utils/buildChoiceOptions";
-import { buildChoiceOptions } from "../utils/buildChoiceOptions";
+import type { ChoiceCandidate } from "@/utils/buildChoiceOptions";
+import { buildChoiceOptions } from "@/utils/buildChoiceOptions";
 
 const BRAND = "#B43D20";
