+++ src/features/exercises/components/SelfRecordButton.tsx
